feat: Add jwt-decode npm dependency for decoding JWT tokens
This commit adds the "jwt-decode" npm dependency to the frontend package.json file. This package is used for decoding JWT tokens. It will be used in the authentication and authorization process of the application.

Also, the "Comment" interface is added to the frontend/src/app/models/comments.models.ts file. This interface defines the structure of a comment object.

Additionally, the "esModuleInterop" option is set to true in the frontend/tsconfig.json file. This option enables interoperability between CommonJS and ES modules.

Lastly, the avisos.service.ts file is modified to remove an unnecessary import statement.

These changes improve the functionality and maintainability of the application.