feat: Add lazy loading for images, update styles, and add Toastr notifications
This commit adds lazy loading for images in the login.component.html file, improving page load performance by loading images only when they are visible on the screen. It also updates the background color and hover effect of the login button in the login.component.css file.

In addition, the ngx-toastr library is added as a dependency in the package.json file, and ToastrModule is imported and configured in the app.module.ts file. This allows for the display of success and error notifications when logging in, enhancing the user experience.

These changes improve the performance, style, and user experience of the application.